structure refactor to allow importing the uat db model
