+++ jobs/auto-analyser-uat/src/auto_analyser/models/__init__.py
