+++ jobs/auto-analyser-uat/src/auto_analyser/models/uat_job_result.py
